LayoutTests/editing/selection/ios/scrolling-with-fixed-selection-does-not-unselect-native-text-field-expected.txt

This test verifies that scrolling while there is a selection in a fixed position container does not unselect text in a focused native text field. This test requires WebKitTestRunner.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS selectedTextInInputField is "Foo bar"
PASS successfullyParsed is true

TEST COMPLETE
Hello
